requirements.txt
django==3.0.5
psycopg2-binary==2.8.4
django-redis==4.11.0

# Restful API
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
drf-yasg==1.17.1
# job queues and job scheduling
celery==4.4.2
flower==0.9.4

# utils
requests==2.23.0
shortuuid==1.0.1
timeago==1.0.13
xpinyin==0.5.6
mistune==0.8.4
BeautifulSoup4==4.8.2

# development only
django-extensions==2.2.9
werkzeug==1.0.1
better-exceptions==0.2.2
ipython==7.13.0

# production
gunicorn
gevent

# alipay sdk
python-alipay-sdk==2.0.1

# test
factory-boy
faker
nose
pytest
pytest-django
pytest-sugar
pytest-cov
